Buttons to add bu , team, and project moves to doc create, vhanges the naming, logo, favicon
