+++ internal/email/templates/subscriber-document-published.html
@@ -8,7 +8,7 @@
     </p>
     <p>
       Cheers,<br />
-      Hermes
+      DocVault
     </p>
   </body>
 </html>
