+++ web/app/components/hermes-logo.hbs
@@ -1,7 +1,7 @@
 {{! @glint-nocheck: not typesafe yet }}
 <div class="hermes-logo logo-container" ...attributes>
   {{!-- <FlightIcon @name="hashicorp" @size="24" /> --}}
-  <img src="images/logo.png" alt="Your Image Alt Text" class="chakra-icon css-1xy62lf logo-img" />
+  <img src="https://avatars.githubusercontent.com/u/7713209?s=48&v=4" alt="Your Image Alt Text" class="chakra-icon css-1xy62lf logo-img" />
   <div class="hermes-logo-divider"></div>
   <div class="hermes-logo-text">
     DocVault
