+++ web/app/components/inputs/product-select/index.hbs
@@ -48,7 +48,7 @@
               class="product-select-selected-value
                 {{unless @selected 'text-color-foreground-faint'}}"
             >
-              {{or @selected "Select a product/area"}}
+              {{or @selected "Select from an existing BU"}}
             </span>
 
             {{#if this.selectedProductAbbreviation}}
