+++ web/app/components/inputs/project-select/index.hbs
@@ -46,7 +46,7 @@
               class="product-select-selected-value
                 {{unless @selected 'text-color-foreground-faint'}}"
             >
-              {{or @selected "Select your Project"}}
+              {{or @selected "Select from an existing Project"}}
             </span>
 
             {{#if this.selectedProductAbbreviation}}
