+++ web/app/components/inputs/team-select/index.hbs
@@ -46,7 +46,7 @@
               class="product-select-selected-value
                 {{unless @selected 'text-color-foreground-faint'}}"
             >
-              {{or @selected "Select your team/pod"}}
+              {{or @selected "Select from an existing team/pod"}}
             </span>
 
             {{#if this.selectedProductAbbreviation}}
