+++ web/app/styles/app.scss
@@ -84,7 +84,7 @@ $background-color_2: #fff;
 }
 
 .logo-img {
-  max-width: 50px; /* Adjust the maximum width as needed */
+  max-width: 30px; /* Adjust the maximum width as needed */
   height: auto; /* Allow the image to scale proportionally */
   padding: 0px; /* Add padding to make it slightly bigger */
   border-radius: 10px; /* Add some border-radius for a rounded look (optional) */
