+++ web/app/templates/application.hbs
@@ -1,5 +1,5 @@
 {{! @glint-nocheck: not typesafe yet }}
-{{page-title "Hermes"}}
+{{page-title "DocVault"}}
 <div id="dialog-container" class="dialog-container"></div>
 
 <Notification />
